% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{dnfsc_sc}
\alias{dnfsc_sc}
\title{Stability-Based Factor Number Estimation}
\usage{
dnfsc_sc(X, K.cand = 1:10, averaging = TRUE)
}
\arguments{
\item{X}{A numeric data matrix matrix of dimensions \eqn{n \times p}.}

\item{K.cand}{An integer vector specifying the candidate set for the number of factors to be evaluated. The default is \code{1:10}.}

\item{averaging}{A logical parameter indicating whether to perform multiple data splits and average the resulting instability measures. If \code{TRUE} (default), the function conducts multiple splits.}
}
\value{
A list containing the results of the estimation, including:
\item{SC1}{Computed values of SC1 for each value in \code{K.cand}.}
\item{SC2}{Computed values of SC2 for each value in \code{K.cand}.}
\item{SC3}{Computed values of SC3 for each value in \code{K.cand}.}
\item{K_hat_SC1}{Estimated number of factors under SC1.}
\item{K_hat_SC2}{Estimated number of factors under SC2.}
\item{K_hat_SC3}{Estimated number of factors under SC3.}
\item{instability}{Vector of loading instabilities for each value in \code{K.cand}.}
}
\description{
This is the core function of the package. It computes the stability-based criteria—SC1, SC2, and SC3—and provides corresponding estimates for the number of factors, as outlined in Lee & Chen (2024+).
}
\examples{
# Generate example dataset
set.seed(123)
X <- example_data.func(scenario =1, signal = 1, n = 1500, p=500)
result <- dnfsc_sc(X, K.cand = 1:10,averaging = TRUE )
}
