% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_example_data.R
\name{example_data.func}
\alias{example_data.func}
\title{Example Data Generator}
\usage{
example_data.func(scenario, signal, n, p)
}
\arguments{
\item{scenario}{An integer specifying the error term distribution: 1 for homogeneous error and 2 for heterogeneous error.}

\item{signal}{An integer indicating the signal strength: 1 for strong signal, 2 for weak signal, and 3 for a mix of equal and mixed-order signals.}

\item{n}{An integer specifying the sample size.}

\item{p}{An integer specifying the number of features.}
}
\value{
A numeric matrix of dimensions \eqn{n \times p} representing the generated data.
}
\description{
This function generates example data according to the settings in Section 4 of Lee & Chen (2024+).
}
\examples{
X <- example_data.func(scenario =1, signal = 1, n = 1500, p=500)
}
